# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SE Project API (Team 8, May 2025)
  description: "Backend API for Team 8's Software Engineering Project (May 2025 Term)\n\
    \nThis backend powers a web-based mobile application designed for senior citizens\
    \ and their family caregivers. The application enhances daily living and care\
    \ coordination through these core modules identified based on user stories:\n\n\
    - Easy-to-use UI: Senior-friendly design with large text, bold colors, and simple\
    \ navigation\n- Medicine Tracker: Prescription management and medication identification\n\
    - Appointments Tracker: Book, view, and get reminders for doctor appointments\n\
    - SOS: Emergency alert system for designated contacts\n- Documents Vault: Secure\
    \ digital repository for medical reports and history\n\n---\n\n## User Story Mapping\n\
    \n| # | User Story (Summary) | API Endpoints |\n|---|----------------------|--------------|\n\
    | 1 | UI accessibility | (Frontend/UI, not API) |\n| 2 | Medicine management |\
    \ `/medicines/`, `/medicines/{medicine_id}`, `/medicines/transcribe` |\n| 3 |\
    \ Appointment management | `/appointments/`, `/appointments/user/{user_id}` |\n\
    | 4 | Emergency contact notification | `/users/{user_id}/sos/trigger`, `/emergency-contacts/`\
    \ |\n| 5 | Document storage | `/documents/`, `/documents/upload`, `/documents/user/{user_id}`\
    \ |\n| 6 | Medication adherence tracking | `/medicines/`, `/medicines/user/{user_id}`\
...
paths:
...
  /api/v1/medicines/transcribe:
    post:
      tags:
      - Medicines
      summary: Transcribe Prescription
      description: 'Transcribe an audio prescription to structured medicine data.
        Requires authentication. Returns structured medicine info or error.


        Uses Gemini 2.5 Flash to take an audio recording of a prescription and convert
        that audio into a JSON that auto-fills the add or update medicine form.


        Supports US2 by enabling users to digitize prescriptions for easier medicine
        management.'
      operationId: transcribe_prescription_api_v1_medicines_transcribe_post
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          type: integer
          title: User Id
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_transcribe_prescription_api_v1_medicines_transcribe_post'
      responses:
        '200':
          description: Prescription transcribed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicineTranscriptionResponse'
        '400':
          description: Invalid audio file.
        '429':
          description: 'RESOURCE_EXHAUSTED: You''ve exceeded the rate limit. See https://ai.google.dev/gemini-api/docs/rate-limits
            for details.'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/medicines/:
    post:
      tags:
      - Medicines
      summary: Create Medicine
      description: 'Create a new medicine record for a user. Requires authentication.
        Clears cache for the user.


        Supports US2 by allowing users to add medicines for tracking and reminders.'
      operationId: create_medicine_api_v1_medicines__post
      parameters:
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Session token for authentication
          title: Session Token
        description: Session token for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicineResponse'
        '201':
          description: Medicine created successfully.
        '400':
          description: Invalid input.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/medicines/user/{user_id}:
    get:
      tags:
      - Medicines
      summary: Get Medicines By User
      description: 'List all medicines for a user. Results are cached for 5 minutes.


        Supports US2 and US6 by providing medicine lists for management and adherence
        tracking.'
      operationId: get_medicines_by_user_api_v1_medicines_user__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the user to get medicines for
          title: User Id
        description: ID of the user to get medicines for
      responses:
        '200':
          description: List of medicines for the user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicineResponse'
                title: Response Get Medicines By User Api V1 Medicines User  User
                  Id  Get
        '404':
          description: User not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/medicines/{medicine_id}:
    get:
      tags:
      - Medicines
      summary: Get Medicine By Id
      description: 'Get a medicine by its ID.


        Supports US2 by letting users and caregivers review specific medicine details.'
      operationId: get_medicine_by_id_api_v1_medicines__medicine_id__get
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the medicine to retrieve
          title: Medicine Id
        description: ID of the medicine to retrieve
      responses:
        '200':
          description: Medicine found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicineResponse'
        '404':
          description: Medicine not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Medicines
      summary: Update Medicine
      description: 'Update an existing medicine record. Requires admin or owner. Clears
        the user''s cache.


        Supports US2 by keeping medicine records up to date for accurate management.'
      operationId: update_medicine_api_v1_medicines__medicine_id__put
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the medicine to update
          title: Medicine Id
        description: ID of the medicine to update
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineUpdate'
      responses:
        '200':
          description: Medicine updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicineResponse'
        '404':
          description: Medicine not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Medicines
      summary: Delete Medicine
      description: 'Delete a medicine record by ID. Requires admin or owner. Clears
        the user''s cache.


        Supports US2 by allowing removal of outdated or incorrect medicines from the
        user''s profile.'
      operationId: delete_medicine_api_v1_medicines__medicine_id__delete
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the medicine to delete
          title: Medicine Id
        description: ID of the medicine to delete
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Medicine deleted successfully.
          content:
            application/json:
              schema: {}
        '404':
          description: Medicine not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_transcribe_prescription_api_v1_medicines_transcribe_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_transcribe_prescription_api_v1_medicines_transcribe_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MedicineCreate:
      properties:
        name:
          type: string
          title: Name
          description: Name of the medicine
          example: Paracetamol
        dosage:
          type: string
          title: Dosage
          description: Dosage information
          example: 500mg
        frequency:
          type: string
          title: Frequency
          description: Frequency of intake
          example: Once a day after dinner
        start_date:
          type: string
          format: date
          title: Start Date
          description: Start date of the medicine course
          example: '2025-06-25'
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
          description: End date of the medicine course (optional)
          example: '2025-07-05'
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Additional notes or instructions (optional)
          example: Take with food
        user_id:
          type: integer
          title: User Id
          description: ID of the user taking the medicine
          example: 1
        doctor_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Doctor Id
          description: ID of the doctor who issued the medicine (optional)
          example: 1
        appointment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Appointment Id
          description: ID of the appointment associated with the medicine (optional)
          example: 1
      type: object
      required:
      - name
      - dosage
      - frequency
      - start_date
      - user_id
      title: MedicineCreate
      description: Schema for creating a new medicine record.
    MedicineResponse:
      properties:
        name:
          type: string
          title: Name
          description: Name of the medicine
          example: Paracetamol
        dosage:
          type: string
          title: Dosage
          description: Dosage information
          example: 500mg
        frequency:
          type: string
          title: Frequency
          description: Frequency of intake
          example: Once a day after dinner
        start_date:
          type: string
          format: date
          title: Start Date
          description: Start date of the medicine course
          example: '2025-06-25'
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
          description: End date of the medicine course (optional)
          example: '2025-07-05'
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Additional notes or instructions (optional)
          example: Take with food
        id:
          type: integer
          title: Id
          description: Unique medicine ID
          example: 1
        user_id:
          type: integer
          title: User Id
          description: ID of the user taking the medicine
          example: 1
        doctor_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Doctor Id
          description: ID of the doctor who issued the medicine (optional)
          example: 1
        appointment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Appointment Id
          description: ID of the appointment associated (optional)
          example: 1
      type: object
      required:
      - name
      - dosage
      - frequency
      - start_date
      - id
      - user_id
      title: MedicineResponse
      description: Schema for returning medicine data to the client.
      example:
        appointment_id: 3
        doctor_id: 2
        dosage: 500mg
        end_date: '2025-07-05'
        frequency: Once a day after dinner
        id: 1
        name: Paracetamol
        notes: Take with food
        start_date: '2025-06-25'
        user_id: 1
    MedicineTranscriptionResponse:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Name of the medicine (optional)
        dosage:
          anyOf:
          - type: string
          - type: 'null'
          title: Dosage
          description: Dosage information (optional)
        frequency:
          anyOf:
          - type: string
          - type: 'null'
          title: Frequency
          description: Frequency of intake (optional)
        start_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Date
          description: Start date of the medicine course (ISO 8601 datetime, optional)
        end_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Date
          description: End date of the medicine course (ISO 8601 datetime, optional)
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Additional notes or instructions (optional)
      type: object
      title: MedicineTranscriptionResponse
      description: Schema for AI-generated medicine transcription from audio.
      example:
        dosage: 500mg
        end_date: '2025-07-05T08:00:00Z'
        frequency: Once a day after dinner
        name: Paracetamol
        notes: Take with food
        start_date: '2025-06-25T08:00:00Z'
    MedicineUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Name of the medicine (optional)
          example: Paracetamol
        dosage:
          anyOf:
          - type: string
          - type: 'null'
          title: Dosage
          description: Dosage information (optional)
          example: 500mg
        frequency:
          anyOf:
          - type: string
          - type: 'null'
          title: Frequency
          description: Frequency of intake (optional)
          example: Once a day after dinner
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
          description: Start date (optional)
          example: '2025-06-25'
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
          description: End date (optional)
          example: '2025-07-05'
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Additional notes (optional)
          example: Take with food
      type: object
      title: MedicineUpdate
      description: Schema for updating a medicine record (partial update).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError